refactor: build & search IVF_HNSW_PQ with new indexer & IVF impl
