+++ rust/lance-index/src/vector/quantizer.rs
@@ -5,10 +5,7 @@ use core::fmt;
 use std::fmt::Debug;
 use std::sync::Arc;
 
-use arrow::array::AsArray;
-use arrow::datatypes::{Float16Type, Float32Type, Float64Type};
 use arrow_array::{Array, ArrayRef, FixedSizeListArray};
-use arrow_schema::DataType;
 use async_trait::async_trait;
 use deepsize::DeepSizeOf;
 use lance_core::{Error, Result};
@@ -23,17 +20,7 @@ use crate::{IndexMetadata, INDEX_METADATA_SCHEMA_KEY};
 
 use super::flat::index::FlatQuantizer;
 use super::pq::ProductQuantizer;
-use super::sq::builder::SQBuildParams;
-use super::sq::storage::SQ_METADATA_KEY;
-use super::SQ_CODE_COLUMN;
-use super::{
-    ivf::storage::IvfModel,
-    sq::{
-        storage::{ScalarQuantizationMetadata, ScalarQuantizationStorage},
-        ScalarQuantizer,
-    },
-    storage::VectorStore,
-};
+use super::{ivf::storage::IvfModel, sq::ScalarQuantizer, storage::VectorStore};
 
 pub trait Quantization: Send + Sync + Debug + DeepSizeOf + Into<Quantizer> {
     type BuildParams: QuantizerBuildParams;
@@ -170,88 +157,6 @@ pub trait QuantizerStorage: Clone + Sized + DeepSizeOf + VectorStore {
         distance_type: DistanceType,
         metadata: &Self::Metadata,
     ) -> Result<Self>;
-}
-
-impl Quantization for ScalarQuantizer {
-    type BuildParams = SQBuildParams;
-    type Metadata = ScalarQuantizationMetadata;
-    type Storage = ScalarQuantizationStorage;
-
-    fn build(data: &dyn Array, _: DistanceType, params: &Self::BuildParams) -> Result<Self> {
-        let fsl = data.as_fixed_size_list_opt().ok_or(Error::Index {
-            message: format!(
-                "SQ builder: input is not a FixedSizeList: {}",
-                data.data_type()
-            ),
-            location: location!(),
-        })?;
-
-        let mut quantizer = Self::new(params.num_bits, fsl.value_length() as usize);
-
-        match fsl.value_type() {
-            DataType::Float16 => {
-                quantizer.update_bounds::<Float16Type>(fsl)?;
-            }
-            DataType::Float32 => {
-                quantizer.update_bounds::<Float32Type>(fsl)?;
-            }
-            DataType::Float64 => {
-                quantizer.update_bounds::<Float64Type>(fsl)?;
-            }
-            _ => {
-                return Err(Error::Index {
-                    message: format!("SQ builder: unsupported data type: {}", fsl.value_type()),
-                    location: location!(),
-                })
-            }
-        }
-
-        Ok(quantizer)
-    }
-
-    fn code_dim(&self) -> usize {
-        self.dim
-    }
-
-    fn column(&self) -> &'static str {
-        SQ_CODE_COLUMN
-    }
-
-    fn quantize(&self, vectors: &dyn Array) -> Result<ArrayRef> {
-        match vectors.as_fixed_size_list().value_type() {
-            DataType::Float16 => self.transform::<Float16Type>(vectors),
-            DataType::Float32 => self.transform::<Float32Type>(vectors),
-            DataType::Float64 => self.transform::<Float64Type>(vectors),
-            value_type => Err(Error::invalid_input(
-                format!("unsupported data type {} for scalar quantizer", value_type),
-                location!(),
-            )),
-        }
-    }
-
-    fn metadata_key() -> &'static str {
-        SQ_METADATA_KEY
-    }
-
-    fn quantization_type() -> QuantizationType {
-        QuantizationType::Scalar
-    }
-
-    fn metadata(&self, _: Option<QuantizationMetadata>) -> Result<serde_json::Value> {
-        Ok(serde_json::to_value(ScalarQuantizationMetadata {
-            dim: self.dim,
-            num_bits: self.num_bits(),
-            bounds: self.bounds(),
-        })?)
-    }
-
-    fn from_metadata(metadata: &Self::Metadata, _: DistanceType) -> Result<Quantizer> {
-        Ok(Quantizer::Scalar(Self::with_bounds(
-            metadata.num_bits,
-            metadata.dim,
-            metadata.bounds.clone(),
-        )))
-    }
 }
 
 /// Loader to load partitioned [VectorStore] from disk.
